refactor(eajs): split embedding-iframe-sdk into script and iframe layers
script/ is the embed.js runtime that runs on customer pages, iframe/ is
the route rendered inside the embed iframe. File moves plus import path
updates in consumers.
