+++ .typedoc/tsconfig.eajs-docs.json
@@ -1,6 +1,6 @@
 {
   "extends": "../tsconfig.base.json",
   "files": [
-    "../frontend/src/metabase/embedding/embedding-iframe-sdk/types/modular-embedding.ts"
+    "../frontend/src/metabase/embedding/embedding-iframe-sdk/script/types/modular-embedding.ts"
   ]
 }
