+++ .typedoc/typedoc.eajs-markdown.config.js
@@ -9,7 +9,7 @@ const eajsMarkdownConfig = {
     "./typedoc-plugin-markdown-prepare-for-embedding.js",
   ],
   entryPoints: [
-    "../frontend/src/metabase/embedding/embedding-iframe-sdk/types/modular-embedding.ts",
+    "../frontend/src/metabase/embedding/embedding-iframe-sdk/script/types/modular-embedding.ts",
   ],
   out: "../docs/embedding/eajs/snippets",
   cleanOutputDir: true,
