+++ frontend/src/metabase/embedding/embedding-iframe-sdk-setup/types.ts
@@ -6,7 +6,7 @@ import type {
   QuestionEmbedOptions,
   SdkIframeEmbedAuthTypeSettings,
   SdkIframeEmbedBaseSettings,
-} from "metabase/embedding/embedding-iframe-sdk/types/embed";
+} from "metabase/embedding/embedding-iframe-sdk/script/types/embed";
 import type {
   BaseRecentItem,
   SdkIframeEmbedSetupTheme,
