+++ frontend/src/metabase/embedding/embedding-iframe-sdk/iframe/hooks/use-sdk-iframe-embed-event-bus.ts
@@ -10,7 +10,7 @@ import type {
   SdkIframeEmbedMessage,
   SdkIframeEmbedSettings,
   SdkIframeEmbedTagMessage,
-} from "../types/embed";
+} from "../../script/types/embed";
 
 import { trackEmbeddedAnalyticsJs } from "./analytics";
