+++ frontend/src/metabase/embedding/embedding-iframe-sdk/iframe/utils/index.ts
