+++ frontend/src/metabase/embedding/embedding-iframe-sdk/iframe/utils/strip-internal-iframe-query-parameters.ts
@@ -1,4 +1,4 @@
-import { EMBED_JS_IFRAME_IDENTIFIER_QUERY_PARAMETER_NAME } from "../constants";
+import { EMBED_JS_IFRAME_IDENTIFIER_QUERY_PARAMETER_NAME } from "../../script/constants";
 
 export const stripInternalIframeQueryParameters = () => {
   const url = new URL(window.location.href);
