+++ frontend/src/metabase/embedding/embedding-iframe-sdk/script/analytics.unit.spec.tsx
@@ -1,6 +1,6 @@
 // Ensure the custom elements are registered
-import "metabase/embedding/embedding-iframe-sdk/embed";
-import type { MetabaseEmbedElement } from "metabase/embedding/embedding-iframe-sdk/embed";
+import "metabase/embedding/embedding-iframe-sdk/script/embed";
+import type { MetabaseEmbedElement } from "metabase/embedding/embedding-iframe-sdk/script/embed";
 
 import { createEmbeddedAnalyticsJsUsage } from "./analytics";
 import type { SdkIframeEmbedBaseSettings } from "./types/embed";
