+++ frontend/src/metabase/embedding/embedding-iframe-sdk/script/constants.ts
