+++ frontend/src/metabase/embedding/embedding-iframe-sdk/script/embed.ts
@@ -1,12 +1,12 @@
 import type { SqlParameterValues } from "embedding-sdk-bundle/types";
 import { MetabaseError, SSO_NOT_ALLOWED } from "embedding-sdk-shared/errors";
 import * as MetabaseErrors from "embedding-sdk-shared/errors";
-import { PLUGIN_EMBED_JS_EE } from "metabase/embedding/embedding-iframe-sdk/plugin";
+import { PLUGIN_EMBED_JS_EE } from "metabase/embedding/embedding-iframe-sdk/script/plugin";
 import type {
   EmbedAuthManager,
   EmbedAuthManagerContext,
-} from "metabase/embedding/embedding-iframe-sdk/types/auth-manager";
-import type { ComponentToAttributes } from "metabase/embedding/embedding-iframe-sdk/types/modular-embedding";
+} from "metabase/embedding/embedding-iframe-sdk/script/types/auth-manager";
+import type { ComponentToAttributes } from "metabase/embedding/embedding-iframe-sdk/script/types/modular-embedding";
 import type { ParameterValues } from "metabase/embedding-sdk/types/dashboard";
 import { decodeJwt } from "metabase/utils/jwt";
