+++ frontend/src/metabase/embedding/embedding-iframe-sdk/script/plugin.ts
@@ -1,4 +1,4 @@
-import type { EmbedAuthManagerConstructor } from "metabase/embedding/embedding-iframe-sdk/types/auth-manager";
+import type { EmbedAuthManagerConstructor } from "metabase/embedding/embedding-iframe-sdk/script/types/auth-manager";
 
 export const PLUGIN_EMBED_JS_EE = {
   // Unjustified type cast. FIXME
