+++ frontend/src/metabase/embedding/embedding-iframe-sdk/script/types/embed.ts
