+++ frontend/src/metabase/embedding/embedding-iframe-sdk/script/types/modular-embedding.ts
