+++ frontend/src/metabase/embedding/embedding-iframe-sdk/script/types/store.ts
